Move helper to base class
